split select into different component
